+++ src/main.rs
@@ -1,5 +1,6 @@
-use std::io::Read;
+use std::{io::Read, str::FromStr};
 
+use anyhow::format_err;
 use structopt::StructOpt;
 
 /// Generates completions based on a word list and a prompt.
@@ -7,12 +8,46 @@ use structopt::StructOpt;
 /// Word list is expected to be provided via stdin, and newline-delimited.
 #[derive(Debug, StructOpt)]
 struct Options {
-    /// Beginning of line to complete against.
-    input: String,
+    /// Pattern to complete against.
+    pattern: String,
 
-    /// Whether matches should ignore case.
-    #[structopt(long, short = "i")]
-    case_insensitive: bool,
+    /// Case matching strategy to use
+    ///
+    /// * auto: Case insensitive if pattern is all lowercase
+    /// * sensitive: Always case sensitive
+    /// * insensitive: Always case insensitive
+    #[structopt(long, default_value = "auto", verbatim_doc_comment)]
+    case: CaseMode,
+}
+
+#[derive(Debug)]
+enum CaseMode {
+    Auto,
+    Sensitive,
+    Insensitive,
+}
+
+impl Default for CaseMode {
+    fn default() -> Self {
+        CaseMode::Auto
+    }
+}
+
+impl FromStr for CaseMode {
+    type Err = anyhow::Error;
+
+    fn from_str(input: &str) -> anyhow::Result<Self> {
+        match input {
+            "auto" => Ok(CaseMode::Auto),
+            "sensitive" => Ok(CaseMode::Sensitive),
+            "insensitive" => Ok(CaseMode::Insensitive),
+
+            _ => Err(format_err!(
+                "Unknown case mode '{}'. Expected one of \
+                 'auto', 'sensitive', or 'insensitive'."
+            )),
+        }
+    }
 }
 
 fn main() {
@@ -24,42 +59,50 @@ fn main() {
         .read_to_string(&mut word_list)
         .unwrap();
 
-    if options.case_insensitive {
-        for word in word_list.lines() {
-            if check_match_insensitive(&options.input, word) {
-                println!("{}", word);
-            }
-        }
-    } else {
-        for word in word_list.lines() {
-            if check_match(&options.input, word) {
-                println!("{}", word);
-            }
+    let case_sensitive = match options.case {
+        CaseMode::Sensitive => true,
+        CaseMode::Insensitive => false,
+        CaseMode::Auto => options
+            .pattern
+            .chars()
+            .any(|char| char.is_ascii_uppercase()),
+    };
+
+    let case_sensitive_matcher = |a, b| a == b;
+    let case_insensitive_matcher = |a: char, b| a.eq_ignore_ascii_case(&b);
+
+    for word in word_list.lines() {
+        if case_sensitive {
+            handle_entry(case_sensitive_matcher, &options.pattern, word);
+        } else {
+            handle_entry(case_insensitive_matcher, &options.pattern, word);
         }
     }
 }
 
-fn check_match(input: &str, word: &str) -> bool {
-    word.starts_with(input)
-}
-
-fn check_match_insensitive(input: &str, word: &str) -> bool {
-    let mut input_chars = input.chars();
+fn handle_entry<F>(matcher: F, pattern: &str, word: &str)
+where
+    F: Fn(char, char) -> bool,
+{
+    let mut pattern_chars = pattern.chars();
     let mut word_chars = word.chars();
 
     loop {
-        match (input_chars.next(), word_chars.next()) {
+        match (pattern_chars.next(), word_chars.next()) {
             (Some(input_char), Some(word_char)) => {
-                if !input_char.eq_ignore_ascii_case(&word_char) {
-                    return false;
+                if !matcher(input_char, word_char) {
+                    return;
                 }
             }
 
             // input is longer than word, which means we cannot match!
-            (Some(_), None) => return false,
+            (Some(_), None) => return,
 
             // input is over, we have matched successfully
-            (None, _) => return true,
+            (None, _) => {
+                println!("{}", word);
+                return;
+            }
         }
     }
 }